+++ legal_knowledge_watch/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import legal_watch_reader